moved utils code from header to source
